api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["kevinwochan <kevinwochan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.119.0"
"aws-cdk.aws-lambda" = "^1.119.0"
"aws-cdk.aws-apigateway" = "^1.119.0"
"aws-cdk.aws-dynamodb" = "^1.119.0"
boto3 = "^1.18.26"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
